refactor(repo C10): extract FTS5 query construction to engine/repository/fts.ts
Moves sanitizeFTS5Term, buildFTS5Query, validateSemanticQuery,
validateLexQuery. searchFTS stays in repository.ts until SearchResult,
getDocid, and getContextForFile move in their respective clusters.

Build: clean. Tests: 970/970 pass.
