+++ engine/repository/index.ts
@@ -11,3 +11,4 @@ export * from "./llm-cache.js";
 export * from "./docid.js";
 export * from "./rerank-queue.js";
 export * from "./chunking.js";
+export * from "./fts.js";
